Left and Right Channel CORRECTION
After testing&listening, I find that the Left Channel and the Right Channel is reversed. This can be confirmed by the channels' subtraction in file "blockProcess.c", that, in the original code, the right earphone still have sound while the left one is muted after I subtract the right channel(the right one should be muted actually). Another way to test it is by changing the algorithm of the last two lines in "DSP_program.c" and the channel that has been changed in code is not actually changed but another channel has.

So I correct this mistake by exchanging them in the code. Now, they are nicely corresponding!

Well, the other parts in the whole project is really perfect and the echo function is effective! Grateful for author's nice job!